==> 1. Merging.ipynb <==
import pandas as pd

cols = ["userID", "movieID", "rating"]
ratings = pd.read_table("movielens-2k/user_ratedmovies.dat", usecols=cols)
ratings.head()

ratings["userID"] = ratings.userID.astype("categorical").cat.codes
ratings["movieID"] = ratings.movieID.astype("categorical").cat.codes

cols = ["id", "title", "year", "imdbPictureURL"]
movies = pd.read_table("movielens-2k/movies.dat", usecols=cols, encoding="latin-1")
movies["movieID"] = movies.id.astype("categorical").cat.codes
movies = movies.drop("id", axis=1)
movies.head()

genres = pd.read_table("movielens-2k/movie_genres.dat")
genres.head()

concat = lambda a: ", ".join(a)

genres = genres.groupby(by="movieID").agg(concat)
genres["movieID"] = genres.movieID.astype("categorical").cat.codes
genres = genres.reset_index()
genres.head()

(ratings
     .merge(movies, on="movieID")
     .merge(genres, on="movieID")
     .dropna(axis=0)
).to_csv("data/movies.csv.bz2", compression="bz2", index=False)
