==> 2. Preprocessing.ipynb <==
import numpy as np
import pandas as pd

cols = ["userID", "movieID", "rating"]
movies = pd.read_csv("data/movies.csv.bz2", usecols=cols)
movies.head()

rating = movies.pivot_table(index="userID", values="rating", columns="movieID", fill_value=0).values
rating

np.savetxt("data/ratings.txt.gz", rating, fmt="%.1f")
